requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-cloud-aiplatform
